+++ src/main/java/com/zjj/project/project/controller/QuestionController.java
@@ -1,7 +1,6 @@
 package com.zjj.project.project.controller;
 
 import com.zjj.project.project.dto.QuestionDTO;
-import com.zjj.project.project.mapper.QuestionMapper;
 import com.zjj.project.project.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -20,7 +19,7 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id, Model model){
+    public String question(@PathVariable(name = "id") Long id, Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         //累加阅读（浏览）数
         questionService.incView(id);
